Fixed issue with logging in
The underlying issue was in the Account_Model class where $querySuccess returns false and exits the if statement if the query doesn't return any results.

This has been removed.

For future reference, this should only be used for testing purposes to allow the program to fail gracefully.

Slight syntactical change in login_controller whilst attempting redesign of if statements attempting to fix issue.